Try to take care of Handler.sendMessage
I have tried to replace Handler.sendMessage(Message) with my created
method.
First of all, I tried to fingure out the msg.what based on backward
analysis from the sendMessage Method.
It is not very perfect. Only if when the msg is created with a constant
value assigned to the msg.what, this analysis will work.