Finished replace the sendMessage method with created myHandler[n] method
but after the replacement, the call graph should be updated.
I think that the swapEdgesOutOf method is not correct.
The edge which will call the myHandler[n] method have been put into the
new call graph;
but the edges happened in the myHandler[n] method have not been
implemented.
I will add it in the constructive process of myHandler[n] method.